reinitialize menu when config has changed
To avoid having to restart the application when the configuration was
changed, an automatic reloading is introduced here. Because it was
simplest to implement, the check is only done on pressing a back button.

So change your config, enter a sub menu, press back and you new config
is active.